apps/ai_core/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pyyaml
chromadb
sentence-transformers
numpy
python-dotenv
